Use correct APIs to get & ping drivers
Also adds offline indicator & handling
