+++ icons.csv
@@ -12,3 +12,4 @@ avatar-default,img/adwaita-icons/status/avatar-default.svg
 dialog-password,img/adwaita-icons/status/dialog-password.svg
 alert,img/octicons/alert.svg
 audio-speakers,img/adwaita-icons/devices/audio-speakers.svg
+network-wireless-offline,img/adwaita-icons/status/network-wireless-offline.svg
