+++ index.html
@@ -54,7 +54,15 @@
 				</svg>
 			</button>
 		</nav>
-		<main class="grid"></main>
+		<main class="grid">
+			<div class="offline-message center error" hidden="">
+				<svg width="128" height="128">
+					<use xlink:href="img/icons.svg#network-wireless-offline">
+				</svg>
+				<br />
+				<strong>You are currently offline</strong>
+			</div>
+		</main>
 		<footer class="relative z-1 color-alt" role="contentinfo">
 			<div id="org-info" itemprop="copyrightHolder" itemtype="http://schema.org/Organization" itemscope="">
 				<h3>
@@ -144,9 +152,10 @@
 			</form>
 		</dialog>
 		<template id="driver-template">
-			<div class="driver card shadow inline-block animation-speed-fast animation-ease-in-out zoomIn" data-driver-id="">
+			<div class="driver card shadow inline-block animation-speed-fast animation-ease-in-out zoomIn" data-imei="" data-driver-id="">
 				<h3 data-field="drivername"></h3>
-				<button type="button" class="btn btn-default" title="Ping Driver">
+				<h4 data-field="vehicle"></h4>
+				<button type="button" class="btn btn-default" data-click="ping" title="Ping Driver">
 					<svg class="current-color icon" xmlns="http://www.w3.org/2000/svg" width="16" height="16" version="1" viewBox="0 0 16 16">
 						<path d="M1 5h2.484l2.97-3L7 2.016v12h-.475L3.485 11H1zm10.875-4a1 1 0 0 0-.594.281l-2 2a1.016 1.016 0 1 0 1.438 1.438l2-2A1 1 0 0 0 11.875 1zM9.781 11a1 1 0 0 0-.5 1.719l2 2a1.016 1.016 0 1 0 1.438-1.438l-2-2A1 1 0 0 0 9.78 11zm.125-4A1.001 1.001 0 1 0 10 9l4.031.031a1 1 0 1 0 0-2L10 7a1 1 0 0 0-.094 0z" overflow="visible"/>
 						<path d="M9 4h1v1H9zm0 7h1v1H9z" overflow="visible"/>
